small typing and wording fixes across the API layer
- get_partitions_base's time_in_s was typed as int | None but is
  fed timestamps (float) by every caller
- NoneForUserWithRealmRoles.__call__ actually returns str | None, not
  None; the job_report endpoint's `user` parameter now matches
- the queries endpoint doesn't use its token_payload dependency beyond
  triggering auth, so name it `_`
- parse_sacct_tres's `values` dict holds both int and float TRES
  counts, annotate it as such
- "Execute a the 'named' query" -> "Execute the 'named' query"

+++ src/slurm_monitor/db/v2/db.py
@@ -497,7 +497,7 @@ class ClusterDB(Database):
 
     async def get_partitions_base(self,
             cluster: str,
-            time_in_s: int | None = None,
+            time_in_s: float | None = None,
             interval_in_s: int = DEFAULT_HISTORY_INTERVAL_IN_S
             ) -> dict[str, list[str]]:
